add new test cases for task service
