+++ app/src/test/java/com/baghdad/tudee/data/service/CategoryServiceImplTest.kt
@@ -2,19 +2,19 @@ package com.baghdad.tudee.data.service
 
 import android.database.sqlite.SQLiteException
 import com.baghdad.tudee.data.database.dao.CategoryDao
-import com.baghdad.tudee.data.service.SampleTaskData.categoryID
-import com.baghdad.tudee.data.service.SampleTaskData.dbError
-import com.baghdad.tudee.data.service.SampleTaskData.duplicateCategoryTitle
-import com.baghdad.tudee.data.service.SampleTaskData.emptyCategoryList
-import com.baghdad.tudee.data.service.SampleTaskData.expectedImageUri
-import com.baghdad.tudee.data.service.SampleTaskData.expectedTitle
-import com.baghdad.tudee.data.service.SampleTaskData.index
-import com.baghdad.tudee.data.service.SampleTaskData.invalidCategoryData
-import com.baghdad.tudee.data.service.SampleTaskData.invalidCategoryId
-import com.baghdad.tudee.data.service.SampleTaskData.nonExistentCategoryId
-import com.baghdad.tudee.data.service.SampleTaskData.oneCategoryExpected
-import com.baghdad.tudee.data.service.SampleTaskData.sampleCategory
-import com.baghdad.tudee.data.service.SampleTaskData.sampleDto
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.categoryID
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.dbError
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.duplicateCategoryTitle
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.emptyCategoryList
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.expectedImageUri
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.expectedTitle
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.index
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.invalidCategoryData
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.invalidCategoryId
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.nonExistentCategoryId
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.oneCategoryExpected
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.sampleCategory
+import com.baghdad.tudee.data.service.shared.TestDummyData.Companion.sampleDto
 import com.baghdad.tudee.domain.exception.DatabaseException
 import io.mockk.coEvery
 import io.mockk.coVerify
